fix(cloud-backend): code-review 10-2 patches — 9 fixes incl beacon misattribution, atomic dismissal, funnel hardening
Agent: Amelia | Phase: BMAD-DEV Step 4 (code-review)
Next: dev-story 10-3 (Conrad/Claudia SOP + drill cadence — currently backlog, needs create-story first)
Blocked: —

Code review (3 adversarial layers): all 6 ACs + 5 decisions cleared. 9 patches
applied + verified (3 browser-verified):
- CC dismissal beacon misattribution → capture escalationId in effect closure
- dwell_focus_ms INTEGER→BIGINT (overflow)
- action_tags jsonb_typeof='array' guard (non-array 500)
- funnel half-open window (boundary parity with report)
- GREATEST() ack-latency clamp (untrusted onboard clock)
- report (N/0) cross-week string + CLI ISO-week anchor
- atomic silent-dismissal insert (WHERE status='unacknowledged', TOCTOU)
- emit on pagehide + visibilitychange (mobile-Safari/bfcache)
4 deferred, 2 dismissed. cloud-backend 168 pass/82.58% cov, ruff+mypy --strict
clean; CC 247 vitest. Story 10-2 → done.

+++ cloud-backend/migrations/versions/0007_escalation_audit.py
@@ -46,7 +46,9 @@ def upgrade() -> None:
         sa.Column("t_event", sa.TIMESTAMP(timezone=True), nullable=False),
         sa.Column("t_fired", sa.TIMESTAMP(timezone=True), nullable=False),
         sa.Column("action_tags", JSONB, nullable=True),
-        sa.Column("dwell_focus_ms", sa.Integer, nullable=True),
+        # BigInteger: dwell is client-supplied; a panel left focused for weeks (or a
+        # buggy/crafted value) would overflow a 32-bit INTEGER and 500 the insert.
+        sa.Column("dwell_focus_ms", sa.BigInteger, nullable=True),
         sa.Column("confidence_score", sa.Float, nullable=True),
         sa.Column("confidence_basis", sa.Text, nullable=True),
         sa.Column("model_versions", JSONB, nullable=True),
